Suggested Histogrammar
I really should bring this up, since we're discussing histograms. Histogrammar would provide interactivity with more than just ROOT: it's supposed to be an exchange format and data analyst's API for aggregation in general. It was originally designed for Spark's parallelism, it's been integrated into ROOT (filling TTrees and providing drawable ROOT objects), and I'm now working on aggregation of quantities on a GPU, for plotting intermediate quantities in processes that run on the GPU (debugging).